streamlit/requirements.txt
numpy
pandas
streamlit
altair
mysql-connector-python
vega_datasets